+++ src/android/LocalNotification.java
@@ -23,10 +23,15 @@
 
 package de.appplant.cordova.plugin.localnotification;
 
+import static android.provider.Settings.ACTION_REQUEST_SCHEDULE_EXACT_ALARM;
+
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.app.KeyguardManager;
 import android.content.Context;
+import android.content.Intent;
+import android.os.Build;
+import android.util.Log;
 import android.util.Pair;
 import android.view.View;
 
@@ -34,6 +39,7 @@ import org.apache.cordova.CallbackContext;
 import org.apache.cordova.CordovaInterface;
 import org.apache.cordova.CordovaPlugin;
 import org.apache.cordova.CordovaWebView;
+import org.apache.cordova.PermissionHelper;
 import org.apache.cordova.PluginResult;
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -42,6 +48,7 @@ import org.json.JSONObject;
 import java.lang.ref.WeakReference;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.function.Predicate;
 
 import de.appplant.cordova.plugin.notification.Manager;
 import de.appplant.cordova.plugin.notification.Notification;
@@ -61,6 +68,9 @@ import static de.appplant.cordova.plugin.notification.Notification.Type.TRIGGERE
 @SuppressWarnings({"Convert2Diamond", "Convert2Lambda"})
 public class LocalNotification extends CordovaPlugin {
 
+    private static final int NOTIFICATION_PERMISSION_CODE = 43334;
+    private static final String NOTIFICATION_PERMISSION = "android.permission.POST_NOTIFICATIONS";
+
     // Reference to the web view for static access
     private static WeakReference<CordovaWebView> webView = null;
 
@@ -72,6 +82,18 @@ public class LocalNotification extends CordovaPlugin {
 
     // Launch details
     private static Pair<Integer, String> launchDetails;
+
+    //Save callback context
+    private CallbackContext callbackContext = null;
+    private JSONArray notificationArguments = null;
+
+    /**
+     * We need this variable because the onResume is being called when user
+     * grants permissions to receive notifications. What a mess.
+     */
+    private boolean requestingNotificationsPermissions = false;
+    private boolean requestingExactAlarmPermission = false;
+    private OSLCNOWarning warning = null;
 
     /**
      * Called after plugin construction and fields have been initialized.
@@ -92,6 +114,13 @@ public class LocalNotification extends CordovaPlugin {
     public void onResume (boolean multitasking) {
         super.onResume(multitasking);
         deviceready();
+
+        if(!requestingNotificationsPermissions && requestingExactAlarmPermission) {
+            requestingExactAlarmPermission = false;
+            onScheduleExactAlarmPermissionResult();
+        }
+
+        requestingNotificationsPermissions = false;
     }
 
     /**
@@ -203,6 +232,14 @@ public class LocalNotification extends CordovaPlugin {
         launchDetails = null;
     }
 
+    @Override
+    public void onRequestPermissionResult(int requestCode, String[] permissions, int[] grantResults) throws JSONException {
+        super.onRequestPermissionResult(requestCode, permissions, grantResults);
+        if(requestCode == NOTIFICATION_PERMISSION_CODE){
+            schedule(notificationArguments, callbackContext);
+        }
+    }
+
     /**
      * Ask if user has enabled permission for local notifications.
      *
@@ -210,8 +247,17 @@ public class LocalNotification extends CordovaPlugin {
      *                JavaScript.
      */
     private void check (CallbackContext command) {
-        boolean allowed = getNotMgr().hasPermission();
-        success(command, allowed);
+        /*
+         * Always true so the code can proceed and schedule the notification.
+         * If false is returned, then no schedule is created.
+         *
+         * Why was this necessary?
+         * By getting this value from 'getNotMgr().hasPermission()', a permission request
+         * popup is presented. However, the user won't have enough time to grand the permission
+         * before the success is sent with 'false' value.
+         * This prevented the notification from being schedule the first time the app opens.
+         */
+        success(command, true);
     }
 
     /**
@@ -262,11 +308,29 @@ public class LocalNotification extends CordovaPlugin {
      *                JavaScript.
      */
     private void schedule (JSONArray toasts, CallbackContext command) {
+        callbackContext = command;
+        notificationArguments = toasts;
+
+        if(Build.VERSION.SDK_INT >= 33
+        && !PermissionHelper.hasPermission(this, NOTIFICATION_PERMISSION)
+        && !requestingNotificationsPermissions){
+            requestingNotificationsPermissions = true;
+            PermissionHelper.requestPermission(this, NOTIFICATION_PERMISSION_CODE, NOTIFICATION_PERMISSION);
+        }
+        else if(hasAnyExactNotification() && !getNotMgr().canScheduleExactAlarms()) {
+            requestScheduleExactAlarmPermission();
+        }
+        else {
+            scheduleWithPermission();
+        }
+    }
+
+    private void scheduleWithPermission(){
         Manager mgr = getNotMgr();
 
-        for (int i = 0; i < toasts.length(); i++) {
-            JSONObject dict    = toasts.optJSONObject(i);
-            Options options    = new Options(dict);
+        for (int i = 0; i < notificationArguments.length(); i++) {
+            JSONObject dict    = notificationArguments.optJSONObject(i);
+            Options options    = new Options(cordova.getActivity(), dict);
             Request request    = new Request(options);
             Notification toast = mgr.schedule(request, TriggerReceiver.class);
 
@@ -275,7 +339,91 @@ public class LocalNotification extends CordovaPlugin {
             }
         }
 
-        check(command);
+        sendScheduleResult(callbackContext, PluginResult.Status.OK, warning);
+        warning = null;
+
+    }
+
+    /**
+     * Checks if there's any notification to be scheduled as exact.
+     *
+     * @return true if there's at least one exact notification. false otherwise.
+     */
+    private boolean hasAnyExactNotification() {
+        return findOptionsWithPredicate(Options::getIsExactNotification);
+    }
+
+    /**
+     * Checks if there's any notification that is mandatory to be scheduled as exact.
+     *
+     * @return true if there's at least one mandatory notification. false otherwise.
+     */
+    private boolean hasAnyMandatoryExactNotification() {
+        return findOptionsWithPredicate(Options::getIsExactMandatory);
+    }
+
+    /**
+     * Applies a predicate for all options from notification arguments
+     *
+     * @param predicate the predicate to apply
+     * @return
+     */
+    private boolean findOptionsWithPredicate(Predicate<Options> predicate) {
+        for (int i = 0; i < notificationArguments.length(); i++) {
+            JSONObject dict = notificationArguments.optJSONObject(i);
+            Options options = new Options(cordova.getActivity(), dict);
+            if(predicate.test(options)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Sets all notification's options to be scheduled as Inexact alarms.
+     */
+    private void setAllNotificationsAsInexact() {
+        for (int i = 0; i < notificationArguments.length(); i++) {
+            JSONObject dict    = notificationArguments.optJSONObject(i);
+            Options options    = new Options(cordova.getActivity(), dict);
+            options.setIsExactNotification(false);
+        }
+    }
+
+    /**
+     * Requests user permission to schedule exact alarms.
+     * The result should be present on onResume method.
+     *
+     */
+    private void requestScheduleExactAlarmPermission() {
+        requestingExactAlarmPermission = true;
+        cordova.getContext().startActivity(new Intent(ACTION_REQUEST_SCHEDULE_EXACT_ALARM));
+    }
+
+    /**
+     * Handles user response to exact alarm permission request.
+     *
+     */
+    private void onScheduleExactAlarmPermissionResult() {
+        Boolean permissionDenied = !getNotMgr().canScheduleExactAlarms();
+
+        if(permissionDenied) {
+            if(hasAnyMandatoryExactNotification()) {
+                /*  permission was denied but there's at least one mandatory exact notification
+                 *  send an error and finish execution
+                 */
+                sendScheduleResult(callbackContext, PluginResult.Status.ERROR, OSLCNOError.EXACT_PERMISSION);
+                return;
+            }
+            /*  permission was denied and there's no mandatory exact notification
+             *  change all notifications to inexact and continue flow normally
+             *  send a warning when execution is done
+             */
+            setAllNotificationsAsInexact();
+            warning = OSLCNOWarning.EXACT_PERMISSION;
+        }
+
+        schedule(notificationArguments, callbackContext);
     }
 
     /**
@@ -290,6 +438,21 @@ public class LocalNotification extends CordovaPlugin {
 
         for (int i = 0; i < updates.length(); i++) {
             JSONObject update  = updates.optJSONObject(i);
+
+            if(!update.has("trigger")){
+                JSONObject trigger = new JSONObject();
+                try{
+                    trigger.put("type", "calendar");
+                    if(!update.get("every").equals("")){
+                        trigger.put("every", update.get("every"));
+                    }
+                    update.put("trigger", trigger);
+                }catch (JSONException e) {
+                    e.printStackTrace();
+                    Log.e("JSON put Exception", "Exception thrown trying to put trigger information into the JSON object holding the update information.");
+                }
+            }
+
             int id             = update.optInt("id", 0);
             Notification toast = mgr.update(id, update, TriggerReceiver.class);
 
@@ -509,6 +672,18 @@ public class LocalNotification extends CordovaPlugin {
     }
 
     /**
+     * Invoke error callback with a string boolean argument.
+     *
+     */
+    private void sendScheduleResult(CallbackContext command,
+                                    PluginResult.Status status,
+                                    OSLCNOPluginMessage message) {
+        command.sendPluginResult(message != null ?
+                new PluginResult(status, message.toJSONObject()) :
+                new PluginResult(status));
+    }
+
+    /**
      * Fire given event on JS side. Does inform all event listeners.
      *
      * @param event The event name.
@@ -563,6 +738,16 @@ public class LocalNotification extends CordovaPlugin {
         }
 
         sendJavascript(js);
+    }
+
+    /**
+     * Current application state.
+     *
+     * @return
+     *      "background" or "foreground"
+     */
+    static String getApplicationState () {
+        return isInForeground() ? "foreground" : "background";
     }
 
     /**